requirement.txt
google-genai == 1.56.0
python-dotenv == 1.2.1
pillow == 12.0.0